+++ src/index.ts
@@ -1,0 +1,4 @@
+import getSchemaForEndpoint from './functions/get-schema-for-endpoint';
+
+export { getSchemaForEndpoint };
+export type { paths, components } from './generated-types';